requirements.txt
python-dotenv==0.19.0
discord==0.16.12